Move knex config to knexMigrator service
